Setting source root
I was wondering why do we scan .class files instead of .java files? If we scan .java files, later on if we open .ozasmt file with AppScan we will be able to view the source code. But by scanning .class file we have to find the related .java files manually ourselves.